--- src/components/LayoutRouter/index.tsx
@@ -1,1 +1,0 @@
-export * from './LayoutRouter';
