+++ src/index.tsx
@@ -1,7 +1,48 @@
+import {
+  Navigate,
+  RouterProvider,
+  createBrowserRouter,
+} from 'react-router-dom';
 import ReactDOM from 'react-dom';
-import { LayoutRouter } from './components/LayoutRouter';
+import { App } from './App';
+import { HomePage } from './pages/HomePage';
+import { TabsPage } from './pages/TabsPage/TabsPage';
+import { NotFoundPage } from './pages/NotFoundPage';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    children: [
+      {
+        index: true,
+        element: <HomePage />,
+      },
+
+      {
+        path: 'home',
+        element: <Navigate to="/" replace />,
+      },
+
+      {
+        path: 'tabs',
+        element: <TabsPage />,
+        children: [
+          {
+            path: ':tabId',
+          },
+        ],
+      },
+
+      {
+        path: '*',
+        element: <NotFoundPage />,
+      },
+    ],
+  },
+]);
 
 ReactDOM.render(
-  <LayoutRouter />,
+  <RouterProvider router={router} />,
   document.getElementById('root'),
 );
